+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/components/project.js
@@ -361,6 +361,11 @@ ProjectChooserView = Backbone.Marionette.ItemView.extend({
         }
     },
 
+    clear: function() {
+        this.set_name();
+        $(this.data_list_selector).empty();
+    },
+
 });